test: refactor codes
